[front/lib/swr] - refactor: move useSearchMembers hook to memberships file
 - Extracted the useSearchMembers hook from user.ts to memberships.ts to better align with its use-case
 - Implemented debouncing for the search term within the useSearchMembers to optimize search functionality
 - Removed redundant imports and cleaned up code in user.ts after moving the search hook
