+++ front/lib/swr/memberships.ts
@@ -1,6 +1,6 @@
 import type { LightWorkspaceType } from "@dust-tt/types";
 import type { PaginationState } from "@tanstack/react-table";
-import { useMemo } from "react";
+import { useEffect, useMemo, useRef, useState } from "react";
 import type { Fetcher } from "swr";
 
 import {
@@ -8,8 +8,10 @@ import {
   fetcher,
   useSWRWithDefaults,
 } from "@app/lib/swr/swr";
+import { debounce } from "@app/lib/utils/debounce";
 import type { GetWorkspaceInvitationsResponseBody } from "@app/pages/api/w/[wId]/invitations";
 import type { GetMembersResponseBody } from "@app/pages/api/w/[wId]/members";
+import type { SearchMembersResponseBody } from "@app/pages/api/w/[wId]/members/search";
 
 export function useMembers({
   workspaceId,
@@ -36,6 +38,55 @@ export function useMembers({
     isMembersError: error,
     mutateMembers: mutate,
     total: data ? data.total : 0,
+  };
+}
+
+export function useSearchMembers({
+  workspaceId,
+  searchTerm,
+  pageIndex,
+  pageSize,
+  disabled,
+}: {
+  workspaceId: string;
+  searchTerm: string;
+  pageIndex: number;
+  pageSize: number;
+  disabled?: boolean;
+}) {
+  const searchMembersFetcher: Fetcher<SearchMembersResponseBody> = fetcher;
+  const debounceHandle = useRef<NodeJS.Timeout | undefined>(undefined);
+  const [debouncedSearchTerm, setDebouncedSearchTerm] = useState(searchTerm);
+
+  useEffect(() => {
+    const debouncedSearch = () => {
+      setDebouncedSearchTerm(searchTerm);
+    };
+
+    debounce(debounceHandle, debouncedSearch, 300);
+  }, [searchTerm]);
+
+  const searchParams = new URLSearchParams({
+    searchTerm: debouncedSearchTerm,
+    orderBy: "name",
+    lastValue: (pageIndex * pageSize).toString(),
+  });
+
+  const { data, error, mutate } = useSWRWithDefaults(
+    `/api/w/${workspaceId}/members/search?${searchParams.toString()}`,
+    searchMembersFetcher,
+    {
+      revalidateOnFocus: false,
+      revalidateOnReconnect: false,
+      disabled,
+    }
+  );
+  return {
+    members: useMemo(() => (data ? data.members : []), [data]),
+    totalMembersCount: data?.total ?? 0,
+    isLoading: !error && !data,
+    isError: !!error,
+    mutateMembers: mutate,
   };
 }
